+++ adapters/web3-react-connector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @blocto/web3-react-connector
 
+## 1.0.3
+
+### Patch Changes
+
+- 4835522: Fix export unreconize type for old typescript
+- Updated dependencies [4835522]
+- Updated dependencies [e742f66]
+  - @blocto/sdk@0.7.1
+
 ## 1.0.2
 
 ### Patch Changes